Fixing url fields.
