+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Entity/Fields/Link/LinkAttribute.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Menu;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Entity\Fields\Link;
 
 use Drupal\Component\Utility\NestedArray;
-use Drupal\Core\Menu\MenuLinkTreeElement;
+use Drupal\link\LinkItemInterface;
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Youshido\GraphQL\Execution\ResolveInfo;
 
@@ -11,24 +11,25 @@ use Youshido\GraphQL\Execution\ResolveInfo;
  * Retrieve specific attributes of a menu link.
  *
  * @GraphQLField(
- *   id = "menu_link_attribute",
+ *   id = "link_item_attribute",
  *   secure = true,
  *   name = "attribute",
  *   type = "String",
- *   parents = {"MenuLink"},
  *   arguments = {
  *     "key" = "String!"
- *   }
+ *   },
+ *   field_types = {"link"},
+ *   deriver = "Drupal\graphql_core\Plugin\Deriver\Fields\EntityFieldPropertyDeriver"
  * )
  */
-class MenuLinkAttribute extends FieldPluginBase {
+class LinkAttribute extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   protected function resolveValues($value, array $args, ResolveInfo $info) {
-    if ($value instanceof MenuLinkTreeElement) {
-      $options = $value->link->getOptions();
+    if ($value instanceof LinkItemInterface) {
+      $options = $value->getUrl()->getOptions();
       yield NestedArray::getValue($options, ['attributes', $args['key']]);
     }
   }
