+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Link/LinkUrl.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Links;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Link;
 
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
-use Drupal\link\LinkItemInterface;
 use Youshido\GraphQL\Execution\ResolveInfo;
+use Drupal\Core\Link;
 
 /**
- * Retrieve a link fields route object.
+ * Retrieve a link's route object.
  *
  * @GraphQLField(
- *   id = "link_item_url",
+ *   id = "link_url",
  *   secure = true,
  *   name = "url",
+ *   description = @Translation("The url of a link."),
  *   type = "Url",
- *   deriver = "Drupal\graphql_core\Plugin\Deriver\Fields\EntityFieldPropertyDeriver",
- *   field_types = {"link"}
+ *   parents = {"Link"}
  * )
  */
 class LinkUrl extends FieldPluginBase {
@@ -24,7 +24,7 @@ class LinkUrl extends FieldPluginBase {
    * {@inheritdoc}
    */
   public function resolveValues($value, array $args, ResolveInfo $info) {
-    if ($value instanceof LinkItemInterface) {
+    if ($value instanceof Link) {
       yield $value->getUrl();
     }
   }
