+++ modules/graphql_core/src/Plugin/GraphQL/Fields/MenuLink/MenuLinkAttribute.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Links;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\MenuLink;
 
 use Drupal\Component\Utility\NestedArray;
-use Drupal\link\LinkItemInterface;
+use Drupal\Core\Menu\MenuLinkTreeElement;
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Youshido\GraphQL\Execution\ResolveInfo;
 
@@ -11,25 +11,24 @@ use Youshido\GraphQL\Execution\ResolveInfo;
  * Retrieve specific attributes of a menu link.
  *
  * @GraphQLField(
- *   id = "link_item_attribute",
+ *   id = "menu_link_attribute",
  *   secure = true,
  *   name = "attribute",
  *   type = "String",
+ *   parents = {"MenuLink"},
  *   arguments = {
  *     "key" = "String!"
- *   },
- *   field_types = {"link"},
- *   deriver = "Drupal\graphql_core\Plugin\Deriver\Fields\EntityFieldPropertyDeriver"
+ *   }
  * )
  */
-class LinkAttribute extends FieldPluginBase {
+class MenuLinkAttribute extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   protected function resolveValues($value, array $args, ResolveInfo $info) {
-    if ($value instanceof LinkItemInterface) {
-      $options = $value->getUrl()->getOptions();
+    if ($value instanceof MenuLinkTreeElement) {
+      $options = $value->link->getOptions();
       yield NestedArray::getValue($options, ['attributes', $args['key']]);
     }
   }
