+++ modules/graphql_core/src/Plugin/GraphQL/Fields/MenuLink/MenuLinkDescription.php
@@ -1,30 +1,30 @@
 <?php
 
-namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Menu;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\MenuLink;
 
 use Drupal\Core\Menu\MenuLinkTreeElement;
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Youshido\GraphQL\Execution\ResolveInfo;
 
 /**
- * Retrieve a menu links route object.
+ * Menu link descriptions.
  *
  * @GraphQLField(
- *   id = "menu_link_url",
+ *   id = "menu_link_description",
  *   secure = true,
- *   name = "url",
- *   type = "Url",
+ *   name = "description",
+ *   type = "String",
  *   parents = {"MenuLink"}
  * )
  */
-class MenuLinkUrl extends FieldPluginBase {
+class MenuLinkDescription extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   public function resolveValues($value, array $args, ResolveInfo $info) {
     if ($value instanceof MenuLinkTreeElement) {
-      yield $value->link->getUrlObject();
+      yield $value->link->getDescription();
     }
   }
