+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Routing/Response/ResponseContent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Routing;
+namespace Drupal\graphql_core\Plugin\GraphQL\Fields\Routing\Response;
 
 use Drupal\graphql\Plugin\GraphQL\Fields\FieldPluginBase;
 use Psr\Http\Message\ResponseInterface;
@@ -8,27 +8,27 @@ use Symfony\Component\HttpFoundation\Response;
 use Youshido\GraphQL\Execution\ResolveInfo;
 
 /**
- * Get the response code of an internal or external request.
+ * Get the response content of an internal or external request.
  *
  * @GraphQLField(
- *   id = "response_code",
+ *   id = "response_content",
  *   secure = true,
- *   name = "code",
- *   type = "Int",
+ *   name = "content",
+ *   type = "String",
  *   parents = {"InternalResponse", "ExternalResponse"}
  * )
  */
-class ResponseCode extends FieldPluginBase {
+class ResponseContent extends FieldPluginBase {
 
   /**
    * {@inheritdoc}
    */
   protected function resolveValues($value, array $args, ResolveInfo $info) {
     if ($value instanceof Response) {
-      yield $value->getStatusCode();
+      yield $value->getContent();
     }
     else if ($value instanceof ResponseInterface) {
-      yield $value->getStatusCode();
+      yield (string) $value->getBody();
     }
   }
