+++ modules/graphql_core/src/Plugin/GraphQL/Fields/Routing/RouteEntity.php
@@ -17,8 +17,6 @@ use Youshido\GraphQL\Execution\ResolveInfo;
 /**
  * Retrieve the current routes entity, if it is an entity route.
  *
- * TODO: Move this to `InternalUrl` (breaking change).
- *
  * @GraphQLField(
  *   id = "route_entity",
  *   secure = true,
